requirements.txt
fastapi
pylint
uvicorn
black
pytest-cov
click
requests
imdbpy
bs4
requests
re